+++ lib/glific/wa_managed_phones.ex
@@ -7,6 +7,7 @@ defmodule Glific.WAManagedPhones do
 
   alias Glific.{
     Contacts,
+    Notifications,
     Providers.Maytapi.ApiClient,
     Repo,
     WAGroup.WAManagedPhone
@@ -183,4 +184,48 @@ defmodule Glific.WAManagedPhones do
   @spec has_any_phones(list()) :: {:ok, list()} | {:error, String.t()}
   defp has_any_phones([]), do: {:error, "No active phones available"}
   defp has_any_phones(wa_managed_phones), do: {:ok, wa_managed_phones}
+
+  @doc """
+  add the phone status
+  """
+  @spec status(String.t(), non_neg_integer()) :: {:ok, WAManagedPhone} | {:error, String.t()}
+  def status(new_status, phone_id) do
+    organization_id = Repo.get_organization_id()
+
+    phone =
+      from(p in WAManagedPhone,
+        where: p.phone_id == ^phone_id
+      )
+      |> Repo.one()
+
+    case phone do
+      nil ->
+        {:error, "Phone ID not found"}
+
+      _phone ->
+        case phone
+             |> WAManagedPhone.changeset(%{status: new_status})
+             |> Repo.update() do
+          {:ok, wa_managed_phone} ->
+            if new_status != "active" do
+              Notifications.create_notification(%{
+                category: "WhatsApp Groups",
+                message:
+                  "Cannot send messages. WhatsApp phone #{wa_managed_phone.phone} is not connected with Maytapi. Current status: #{wa_managed_phone.status}",
+                severity: Notifications.types().critical,
+                organization_id: organization_id,
+                entity: %{
+                  phone: wa_managed_phone.phone,
+                  status: new_status
+                }
+              })
+            end
+
+            {:ok, wa_managed_phone}
+
+          {:error, _changeset} ->
+            {:error, "Failed to update status"}
+        end
+    end
+  end
 end
