+++ lib/glific_web/providers/maytapi/controller/status_controller.ex
@@ -2,7 +2,7 @@ defmodule GlificWeb.Providers.Maytapi.Controllers.StatusController do
   @moduledoc false
 
   use GlificWeb, :controller
-  alias Glific.Providers.Maytapi.ApiClient
+  alias Glific.WAManagedPhones
 
   @doc false
   @spec handler(Plug.Conn.t(), map(), String.t()) :: Plug.Conn.t()
@@ -15,7 +15,7 @@ defmodule GlificWeb.Providers.Maytapi.Controllers.StatusController do
   @doc false
   @spec status(Plug.Conn.t(), map()) :: Plug.Conn.t()
   def status(conn, %{"status" => status, "phone_id" => phone_id} = params) do
-    ApiClient.status(status, phone_id)
+    WAManagedPhones.status(status, phone_id)
 
     handler(conn, params, "status handler")
   end
